add shit to eval command

+++ src/commands/owner/eval.ts
@@ -1,5 +1,5 @@
+import chalk from 'chalk';
 import { exec } from 'child_process';
-import { Command } from 'discord-akairo';
 import { MessageEmbed } from 'discord.js';
 import { promisify } from 'util';
 import { inspect } from 'util';
@@ -13,21 +13,9 @@ export default class evaluate extends BotCommand {
         super('eval', {
             aliases: ['eval', 'ev', 'exec'],
             args: [
-                {
-                    id: 'codetoeval',
-                    type: 'string',
-                    match: 'rest'
-                },
-                {
-                    id: "silent",
-                    match: 'flag',
-                    flag: '--silent',
-                },
-                {
-                    id: 'sudo',
-                    match: 'flag',
-                    flag: '--sudo'
-                }
+                { id: 'codetoeval', type: 'string', match: 'rest' },
+                { id: "silent", match: 'flag', flag: '--silent', },
+                { id: 'sudo', match: 'flag', flag: '--sudo' }
             ],
             ownerOnly: true,
             channel: 'guild'
@@ -42,7 +30,7 @@ export default class evaluate extends BotCommand {
             if (args.codetoeval.includes(`env`)) {
                 return message.channel.send(`no env`)
             }
-            
+
             if (args.codetoeval.includes(`message.channel.delete`)) {
                 return message.channel.send(`Are you IRONM00N?`)
             }
@@ -53,45 +41,60 @@ export default class evaluate extends BotCommand {
                 return message.channel.send(`This would be blocked by smooth brain protection, but BushBot has a license`)
             }
 
-	    let embed = new MessageEmbed()
-	    let guild = message.guild
-	    let client = this.client
-	    let channel = message.channel
-            let output = await eval(args.codetoeval)
+            async function send(thingToSend: string) {
+                message.channel.send(thingToSend)
+            }
 
-            const tokencheck = inspect(output)
+            let guild = message.guild
+            let client = this.client
+            let channel = message.channel
+            let embed = new MessageEmbed()
+            let user = message.author
+            let member = message.member
+            let botUser = this.client.user
+            let botMember = message.guild.me
+
+            let output = await eval(args.codetoeval)
 
             if (inspect(output).includes(process.env["token"])) {
                 return message.channel.send(`Message containing token wasn't sent.`)
             }
 
+            //im going to make something that disables eval embed in specific channels in the database later
+            if (message.guild.id == `794610828317032458` && message.channel.id != `834878498941829181`) {
+                if (args.codetoeval.includes('message.delete')) {
+                    return
+                }
+                return message.react(`<:success:838816341007269908>`)
+            }
+
             if (!args.silent && !args.codetoeval.includes("message.channel.delete()")) {
-                const evaloutputembed = new MessageEmbed()
+                const evalOutputEmbed = new MessageEmbed()
                     .setTitle('Evaluated Code')
                     .addField(`:inbox_tray: **Input**`, `\`\`\`js\n${args.codetoeval}\`\`\``)
 
                 if (inspect(output, { depth: 0 }).length > 1000) {
-                    await evaloutputembed.addField(`:outbox_tray: **Output**`, await utils.haste(inspect(output)))
+                    await evalOutputEmbed.addField(`:outbox_tray: **Output**`, await utils.haste(inspect(output)))
                 }
                 else {
-                    evaloutputembed.addField(`:outbox_tray: **Output**`, `\`\`\`js\n${inspect(output, { depth: 0 })}\`\`\``)
+                    evalOutputEmbed.addField(`:outbox_tray: **Output**`, `\`\`\`js\n${inspect(output, { depth: 0 })}\`\`\``)
                 }
 
-                await message.channel.send(evaloutputembed)
+                await message.channel.send(evalOutputEmbed)
             }
             if (args.silent) {
                 if (args.codetoeval.includes('message.delete')) {
                     return
                 }
-                message.react(`<:green_check:796548440266899526>`)
+                message.react(`<:success:838816341007269908>`)
             }
+
         }
         catch (err) {
             try { utils.errorhandling(err, message) }
             catch (err) {
                 utils.errorchannelsend(err)
             }
-
         }
     }
 }
